chore: bundle cfn-guard (#6)
This adds functionality around bundling the cfn-guard binary

I've added two tasks:
1. `update-guard`: this task can be run to query the latest guard
release from GitHub and update the `guard-version.json` file. There is
also a corresponding workflow that will run on a schedule to run this
task and create a PR if there is a new version
2. `bundle-guard`: this task takes the release_id from
`guard-version.json` and downloads the guard binary and puts it under
`bin`

I'm not committing the binaries because we can just bundle the binary
during the build.

Fixes #

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>