chore: use `version` and `ruleIds` properties (#79)
Populating the `version` and `ruleIds` property

The `ruleIds` property will be populated with a shorter version of the
rule name (e.g. `ct-s3-pr-1` becomes `s3pr1`) in order to cut down on
the total length.

This required an upgrade to `2.76` of cdk which also changed the synth
behavior to not throw an error. I updated all the tests to assert on
exitCode instead.

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>