feat: upgrade cfn-guard to 3.0.0
This PR upgrades `cfn-guard` to the latest major version [3.0.0](https://github.com/aws-cloudformation/cloudformation-guard/releases/tag/3.0.0)

The major feature that we are taking advantage of is the new
`--structured` flag which combines all the output into a json parseable
list. I tested this out in a medium sized project and it reduced the
time to execute the plugin from ~2 minutes ~2 seconds!